Fixes warnings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@62841 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/public/information/meshdistance.cc
@@ -79,7 +79,7 @@ float AimsMeshDistance( const AimsData< float > & refMap,
 					break;
 				}
 				else {
-					floorPixCoords[d] = floor( pixCoord );
+					floorPixCoords[d] = (int16_t)floor( pixCoord );
 					Weights[d] = 1 - ( pixCoord - floorPixCoords[d] );
 				}
 			}
@@ -108,7 +108,7 @@ float AimsMeshDistance( const AimsData< float > & refMap,
 					break;
 				}
 				else {
-					floorPixCoords[d] = floor( pixCoord );
+					floorPixCoords[d] = (int16_t)floor( pixCoord );
 					Weights[d] = 1 - ( pixCoord - floorPixCoords[d] );
 				}
 			}
